Refactor Code - merge 충돌 때문에 SpecsForPurchase 를 주석처리 했습니다.
