+++ payment-api/payment-application/src/main/java/kr/flab/movieon/payment/application/PaymentFacade.java
@@ -4,8 +4,6 @@ import java.math.BigDecimal;
 import kr.flab.movieon.payment.domain.Payment;
 import kr.flab.movieon.payment.domain.PaymentApprovalCommand;
 import kr.flab.movieon.payment.domain.PaymentProcessor;
-import kr.flab.movieon.payment.domain.PaymentRepository;
-import kr.flab.movieon.payment.domain.exception.InvalidPaymentCommandException;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,14 +12,11 @@ import org.springframework.transaction.annotation.Transactional;
 public class PaymentFacade {
 
     private final PaymentProcessor paymentProcessor;
-    private final PaymentRepository paymentRepository;
     private final ApplicationEventPublisher publisher;
 
     public PaymentFacade(PaymentProcessor paymentProcessor,
-        PaymentRepository paymentRepository,
         ApplicationEventPublisher publisher) {
         this.paymentProcessor = paymentProcessor;
-        this.paymentRepository = paymentRepository;
         this.publisher = publisher;
     }
 
@@ -34,11 +29,7 @@ public class PaymentFacade {
             .pgToken(request.getPgToken())
             .build();
 
-        paymentProcessor.pay(command);
-
-        var payment = paymentRepository.findByOrderId(command.getOrderId())
-            .orElseThrow(() -> new InvalidPaymentCommandException(
-                "The payment cannot be completed, Failed to complete payment."));
+        var payment = paymentProcessor.pay(command);
 
         payment.pollAllEvents().forEach(publisher::publishEvent);
     }
